# v6: more efficient evaluation