--- a/v6/config_cache.go
+++ b/v6/config_cache.go
@@ -1,5 +1,7 @@
 package configcat
 
+import "fmt"
+
 // ConfigCache is a cache API used to make custom cache implementations.
 type ConfigCache interface {
 	// get reads the configuration from the cache.
@@ -8,22 +10,46 @@ type ConfigCache interface {
 	Set(key string, value string) error
 }
 
-type inMemoryConfigCache struct {
-	store map[string]string
+type configCache interface {
+	// get reads the configuration from the cache.
+	get(key string) (*config, error)
+	// set writes the configuration into the cache.
+	set(key string, value *config) error
 }
 
-// newInMemoryConfigCache creates an in-memory cache implementation used to store the fetched configurations.
-func newInMemoryConfigCache() *inMemoryConfigCache {
-	return &inMemoryConfigCache{store: make(map[string]string)}
-}
+type inMemoryConfigCache map[string]*config
 
 // get reads the configuration from the cache.
-func (cache *inMemoryConfigCache) Get(key string) (string, error) {
-	return cache.store[key], nil
+func (cache inMemoryConfigCache) get(key string) (*config, error) {
+	return cache[key], nil
 }
 
 // set writes the configuration into the cache.
-func (cache *inMemoryConfigCache) Set(key string, value string) error {
-	cache.store[key] = value
+func (cache inMemoryConfigCache) set(key string, value *config) error {
+	cache[key] = value
 	return nil
+}
+
+func adaptCache(c ConfigCache) configCache {
+	return &cacheAdaptor{c}
+}
+
+type cacheAdaptor struct {
+	c ConfigCache
+}
+
+func (c *cacheAdaptor) get(key string) (*config, error) {
+	val, err := c.c.Get(key)
+	if err != nil {
+		return nil, err
+	}
+	conf, err := parseConfig([]byte(val))
+	if err != nil {
+		return nil, fmt.Errorf("cannot parse config from cache: %v", err)
+	}
+	return conf, nil
+}
+
+func (c *cacheAdaptor) set(key string, value *config) error {
+	return c.c.Set(key, value.body())
 }