--- a/v6/config_parser.go
+++ b/v6/config_parser.go
@@ -1,7 +1,10 @@
 package configcat
 
 import (
+	"crypto/sha1"
+	"encoding/binary"
 	"encoding/json"
+	"fmt"
 	"strings"
 )
 
@@ -13,91 +16,72 @@ func (p *parseError) Error() string {
 	return p.msg
 }
 
-type configParser struct {
-	evaluator *rolloutEvaluator
-	logger    Logger
+type config struct {
+	jsonBody string
+	root     *rootNode
 }
 
-func newParser(logger Logger) *configParser {
-	evaluator := newRolloutEvaluator(logger)
-	return &configParser{evaluator: evaluator, logger: logger}
-}
-
-func (parser *configParser) parse(jsonBody string, key string, user *User) (interface{}, error) {
-	result, _, err := parser.parseInternal(jsonBody, key, user)
-	return result, err
-}
-
-func (parser *configParser) parseVariationId(jsonBody string, key string, user *User) (string, error) {
-	_, variationId, err := parser.parseInternal(jsonBody, key, user)
-	return variationId, err
-}
-
-func (parser *configParser) getAllKeys(jsonBody string) ([]string, error) {
-	entries, err := parser.getEntries(jsonBody)
-	if err != nil {
+func parseConfig(jsonBody []byte) (*config, error) {
+	var root rootNode
+	if err := json.Unmarshal([]byte(jsonBody), &root); err != nil {
 		return nil, err
 	}
-
-	keys := make([]string, len(entries))
-	i := 0
-	for k := range entries {
-		keys[i] = k
-		i++
-	}
-
-	return keys, nil
+	return &config{
+		jsonBody: string(jsonBody),
+		root:     &root,
+	}, nil
 }
 
-func (parser *configParser) parseKeyValue(jsonBody string, variationId string) (string, interface{}, error) {
-	entries, err := parser.getEntries(jsonBody)
-	if err != nil {
-		return "", nil, &parseError{"JSON parsing failed. " + err.Error() + "."}
+func (c *config) body() string {
+	if c == nil {
+		return ""
 	}
+	return c.jsonBody
+}
 
-	for key, entry := range entries {
+func (conf *config) getKeyAndValueForVariation(variationId string) (string, interface{}) {
+	for key, entry := range conf.root.Entries {
 		if entry.VariationID == variationId {
-			return key, entry.Value, nil
+			return key, entry.Value
 		}
 		for _, rule := range entry.RolloutRules {
 			if rule.VariationID == variationId {
-				return key, rule.Value, nil
+				return key, rule.Value
 			}
 		}
 		for _, rule := range entry.PercentageRules {
 			if rule.VariationID == variationId {
-				return key, rule.Value, nil
+				return key, rule.Value
 			}
 		}
 	}
-
-	return "", nil, &parseError{"JSON parsing failed."}
+	return "", nil
 }
 
-func (parser *configParser) parseInternal(jsonBody string, key string, user *User) (interface{}, string, error) {
+func (conf *config) getAllKeys() []string {
+	keys := make([]string, 0, len(conf.root.Entries))
+	for k := range conf.root.Entries {
+		keys = append(keys, k)
+	}
+	return keys
+}
+
+func (conf *config) getValueAndVariationId(logger Logger, key string, user *User) (interface{}, string, error) {
 	if len(key) == 0 {
-		panic("Key cannot be empty")
+		return nil, "", fmt.Errorf("key cannot be empty")
 	}
-
-	entries, err := parser.getEntries(jsonBody)
-	if err != nil {
-		return nil, "", &parseError{"JSON parsing failed. " + err.Error() + "."}
+	if conf == nil {
+		return nil, "", fmt.Errorf("no configuration available")
 	}
-
-	entryNode := entries[key]
+	entryNode := conf.root.Entries[key]
 	if entryNode == nil {
-		keys := make([]string, len(entries))
-		i := 0
-		for k := range entries {
-			keys[i] = k
-			i++
+		return nil, "", &parseError{
+			"Value not found for key " + key +
+				". Here are the available keys: " + strings.Join(conf.getAllKeys(), ", "),
 		}
-
-		return nil, "", &parseError{"Value not found for key " + key +
-			". Here are the available keys: " + strings.Join(keys, ", ")}
 	}
 
-	parsed, variationId := parser.evaluator.evaluate(entryNode, key, user)
+	parsed, variationId := conf.evaluate(logger, entryNode, key, user)
 	if parsed == nil {
 		return nil, "", &parseError{"Null evaluated for key " + key + "."}
 	}
@@ -105,20 +89,79 @@ func (parser *configParser) parseInternal(jsonBody string, key string, user *Use
 	return parsed, variationId, nil
 }
 
-func (parser *configParser) getEntries(jsonBody string) (map[string]*entry, error) {
-	root, err := parser.deserialize(jsonBody)
-	if err != nil {
-		return nil, err
-	}
-	return root.Entries, nil
-}
+func (conf *config) evaluate(logger Logger, node *entry, key string, user *User) (interface{}, string) {
+	logger.Infof("Evaluating GetValue(%s).", key)
 
-func (parser *configParser) deserialize(jsonBody string) (*rootNode, error) {
-	var root rootNode
-	if err := json.Unmarshal([]byte(jsonBody), &root); err != nil {
-		return nil, err
+	if user == nil {
+		if len(node.RolloutRules) > 0 || len(node.PercentageRules) > 0 {
+			logger.Warnln("Evaluating GetValue(" + key + "). UserObject missing! You should pass a " +
+				"UserObject to GetValueForUser() in order to make targeting work properly. " +
+				"Read more: https://configcat.com/docs/advanced/user-object.")
+		}
+
+		result := node.Value
+		logger.Infof("Returning %v.", result)
+		return result, node.VariationID
 	}
-	return &root, nil
+
+	logger.Infof("User object: %v", user)
+
+	for _, rule := range node.RolloutRules {
+		if rule == nil {
+			continue
+		}
+		matched, err := matchRolloutRule(rule, user)
+		if err != nil {
+			logger.Infof("Evaluating rule: [%s:%s] [%s] [%s] => SKIP rule. Validation error: %v",
+				rule.ComparisonAttribute,
+				user,
+				rule.Comparator,
+				rule.Value,
+				err,
+			)
+			continue
+		}
+		if matched {
+			logger.Infof("Evaluating rule: [%s:%s] [%s] [%s] => match, returning: %v",
+				rule.ComparisonAttribute,
+				user,
+				rule.Comparator,
+				rule.ComparisonValue,
+				rule.Value,
+			)
+			return rule.Value, rule.VariationID
+		}
+		logger.Infof("Evaluating rule: [%s:%s] [%s] [%s] => no match",
+			rule.ComparisonAttribute,
+			user,
+			rule.Comparator,
+			rule.Value,
+		)
+	}
+
+	if len(node.PercentageRules) > 0 {
+		sum := sha1.Sum([]byte(key + user.identifier))
+		// Treat the first 4 bytes as a number, then knock
+		// of the last 4 bits. This is equivalent to turning the
+		// entire sum into hex, then decoding the first 7 digits.
+		num := int64(binary.BigEndian.Uint32(sum[:4]))
+		num >>= 4
+
+		scaled := num % 100
+		bucket := int64(0)
+		for _, rule := range node.PercentageRules {
+			bucket += rule.Percentage
+			if scaled < bucket {
+				result := rule.Value
+				logger.Infof("Evaluating %% options. Returning %s", result)
+				return result, rule.VariationID
+			}
+		}
+	}
+
+	result := node.Value
+	logger.Infof("Returning %v.", result)
+	return result, node.VariationID
 }
 
 type rootNode struct {