--- a/v6/manual_polling_policy.go
+++ b/v6/manual_polling_policy.go
@@ -24,7 +24,7 @@ func ManualPoll() RefreshMode {
 // newManualPollingPolicy initializes a new manualPollingPolicy.
 func newManualPollingPolicy(
 	configFetcher configProvider,
-	cache ConfigCache,
+	cache configCache,
 	logger Logger,
 	sdkKey string) *manualPollingPolicy {