--- a/v6/rollout_integration_test.go
+++ b/v6/rollout_integration_test.go
@@ -23,6 +23,22 @@ type integrationTest struct {
 	kind     int
 }
 
+func BenchmarkGetValue(b *testing.B) {
+	b.ReportAllocs()
+	logger := DefaultLogger(LogLevelError)
+	client := NewCustomClient(integrationTests[0].sdkKey, ClientConfig{Logger: logger, Mode: LazyLoad(120, true)})
+	client.Refresh()
+	defer client.Close()
+	b.ResetTimer()
+	val := client.GetValueForUser("bool30TrueAdvancedRules", nil, nil)
+	if val != true {
+		b.Fatalf("unexpected result %#v", val)
+	}
+	for i := 0; i < b.N; i++ {
+		client.GetValueForUser("bool30TrueAdvancedRules", nil, nil)
+	}
+}
+
 var integrationTests = []integrationTest{{
 	sdkKey:   "PKDVCLf-Hq-h-kCzMp-L7Q/psuH7BGHoUmdONrzzUOY7A",
 	fileName: "testmatrix.csv",