… v6/auto_polling_policy_test.go
… v6/config_fetcher_test.go
… v6/config_parser_test.go
… v6/configcat_client_async_test.go
… v6/configcat_client_test.go
… v6/lazy_loading_policy_test.go
… v6/manual_polling_policy_test.go
… v6/refresh_policy.go
… v6/rollout_evaluator.go