Rather than parse the JSON from a string every time we want to evaluate
something, we parse it once only when we fetch it.

This required a bit of internal refactoring (the external API should remain exactly the same):
- change `configParser` into the immutable `config` type
- use an internal version of `ConfigCache` which caches the already-parsed configuration
- rename `parseJson` to `getValue` (and other related renames - these methods don't parse the JSON any more).

Also added a benchmark to check that this speeds things up by a couple of orders of magnitude:

```
benchmark               old ns/op     new ns/op     delta
BenchmarkGetValue-8     82673         752           -99.09%

benchmark               old allocs     new allocs     delta
BenchmarkGetValue-8     346            10             -97.11%

benchmark               old bytes     new bytes     delta
BenchmarkGetValue-8     13557         504           -96.28%
```